Backported jquery.jplayer.js to 2.0.21 and other bug fixes and enhancements
Had to revere jquery.jplayer.js back to version based on 2.0.21 to fix
duration bugs, one of the latest master checkins of a newer version
appears to be causing problems with duration and currenttime updates,
need to dig into and find the cause. Also added more rtmp source parsing
robustness in JplayerRtmp.as, added some more details to the
RTMP_SUPPORT_README, and updated the other ActionScript files to the
latest 2.0.26 versions. Only the core jquery.jplayer.js is now out of
sync - hope to fix this soon and merge the latest changes and then issue
another Pull request, probably after 2.1 is released.