Fixes for rapid successive setMedia and new logging
Add some workarounds (possibly a little kludgy) so that if setMedia is
called in rapid succession while the player is currently playing,
ConnectionManager can properly shut things down and handle closed
connections, and short-circuit any active connection attempts (what a
pain!) and edge case, but easy to run into.

Also added in support for a new external helper AIR app (to be
committed soon) to watch/save debug and log output. Right now much of
the logging and debug events are disabled temporarily while
troubleshooting another potential glitch with LIVE rtmp streams.

Also, this version still reports older values, so jquery.jplayer.js in
the last commit looks for these older versions of the player, and not
the most recent versions. Will sync/update the AS version #'s reported
to match the latest master 2.1 release in the next commit or two...