Update actionscript/happyworm/jPlayer/JplayerRtmp.as
There is an issue with the RTMP video playback. This issue is that all videos are set to the incorrect size and do not update with the metaData dimensions from the video.
This fix removes the "myStatus.metaDataReady = true;" line so the flag will stay "false" and allow the video's size to be set in the "onMetaDataHandler" function. The incorrect sizing of the RTMP video can be seen in the RTMP Demo at the following URL: "http://jplayer.org/latest/demo-01-video-supplied-rtmpv/"

The change is on line 366 (this line is commented out). I could not get the file to only show the difference of this one line.