Added jQuery UI Sortable support for jPlayer Playlists add-on
Added new demo page without minified resources to demonstrate UI
sortable functionality. This modification inserts JSON script tags for
each playlist item so that it can rebuild the playlist from the DOM.
Main benefit here is that playback is not interrupted when user makes
changes to the playlist order. Upon playing, or skipping to previous or
next track, jPlayerPlaylist now evaluates the DOM and rebuilds the
playlist object, before playing the correct index. It supports looping
and it follows the same philosophy as enableRemoveControls.